Remove foldTree, add alternative implementation
My motivation for creating this tic-tac-toe project is to use it in a
Miso project.  Miso depends on ghcjs.  Using ghcjs forces me to use
stack lts-9.21 in the Miso project.  The Data.Tree module from lts-9.21
(containers-0.5.7.1) doesn't have the foldTree function.  Prior to this
commit, I used the foldTree function (from containers-0.6.0.1), but that
doesn't work for the Miso project.  So I removed the use of foldTree and
wrote some new code that accomplishes the same thing without the use
of foldTree.