Tests: make arbitrary Ints when creating StdGens
When using the initGameState or makeSampleGS functions, you must provide
an Int that will be used to create a StdGen in the GameState.  This
StdGen gives the AI the ability to select moves at random.

In the tests, use QuickCheck to generate arbitrary Ints whenever an
StdGen seed is needed.