Improve doComputerMove AI in early game
Prior to this commit, if the Human played the first move of the game in
a corner cell, doComputerMove would play into the opposite corner.  That
is a bad strategy.  The best move in that case is to play in the center
cell.  That is now what doComputerMove will do in that scenario.