Test that AI actually starts in corners and center
Add a test to check that the AI will actually play into the corners or
the center as the first move of the game.  Because the AI will randomly
choose which of those cells to play into, special care had to be taken
to make sure that the test actually checks that the AI really will
choose each of those options given enough opportunities to do so.